requirements.txt
moviepy==1.0.3
opencv_python==4.8.0.76
opencv_python_headless==4.9.0.80
Pillow==10.0.0
Pillow==10.2.0
streamlit==1.18.1
ultralytics==8.1.20
